pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scribe-agent"
version = "0.1.0"
description = "Scribe: watch .txt files → Llama Stack + MCP + Minimax → GitHub issues"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "llama-stack-client>=0.3.0",
    "GitPython>=3.1.40",
    "pydantic-settings>=2.2.0",
]

[project.scripts]
scribe-agent = "scribe_agent.main:main"

[tool.setuptools.packages.find]
where = ["src"]
